requirements.txt
django
djangorestframework
django-filter
sympy
numpy
requests
matplotlib